+++ src/test/java/com/reason/ide/search/reference/FindLIdentUsagesRMLTest.java
@@ -78,7 +78,7 @@ public class FindLIdentUsagesRMLTest extends ORBasePlatformTestCase {
         List<UsageInfo> usages = findUsages("A.re");
         assertSize(1, usages);
         UsageInfo usageInfo = usages.get(0);
-        assertEquals("A.x.f2", ((RPsiQualifiedPathElement) usageInfo.getElement().getParent()).getQualifiedName());
+        assertEquals("A.x.f2", ((RPsiQualifiedPathElement) usageInfo.getValues().getParent()).getQualifiedName());
     }
 */
 }