+++ src/test/java/com/reason/lang/reason/ModuleParsingTest.java
@@ -46,7 +46,7 @@ public class ModuleParsingTest extends RmlParsingTestCase {
         RPsiInnerModule module = first(moduleExpressions(parseCode("module type Intf = { let x: bool; };")));
 
         assertEquals("Intf", module.getName());
-        assertTrue(module.isInterface());
+        assertTrue(module.isModuleType());
         assertInstanceOf(module.getBody(), RPsiModuleBinding.class);
 
     }
@@ -69,7 +69,7 @@ public class ModuleParsingTest extends RmlParsingTestCase {
 
         assertEquals(1, expressions(file).size());
         assertEquals("Router", module.getName());
-        assertEquals("{ let watchUrl: (url => unit) => watcherID; }", module.getModuleType().getText());
+        assertEquals("{ let watchUrl: (url => unit) => watcherID; }", module.getModuleSignature().getText());
         assertNull(PsiTreeUtil.findChildOfType(file, RPsiScopedExpr.class));
         assertNull(module.getBody());
         RPsiLet let = PsiTreeUtil.findChildOfType(file, RPsiLet.class);
@@ -81,8 +81,8 @@ public class ModuleParsingTest extends RmlParsingTestCase {
         RPsiInnerModule e = firstOfType(parseCode("module M: MType = { type t = int; };"), RPsiInnerModule.class);
 
         assertEquals("M", e.getName());
-        assertEquals("MType", e.getModuleType().getText());
-        assertEquals(myTypes.A_MODULE_NAME, e.getModuleType().getFirstChild().getNode().getElementType());
+        assertEquals("MType", e.getModuleSignature().getText());
+        assertEquals(myTypes.A_MODULE_NAME, e.getModuleSignature().getFirstChild().getNode().getElementType());
         assertEquals("{ type t = int; }", e.getBody().getText());
     }
 
@@ -91,8 +91,8 @@ public class ModuleParsingTest extends RmlParsingTestCase {
         RPsiInnerModule e = firstOfType(parseCode("module M: I with type t = X.t = {};"), RPsiInnerModule.class);
 
         assertEquals("M", e.getName());
-        assertEquals("I", e.getModuleType().getText());
-        assertEquals(myTypes.A_MODULE_NAME, e.getModuleType().getFirstChild().getNode().getElementType());
+        assertEquals("I", e.getModuleSignature().getText());
+        assertEquals(myTypes.A_MODULE_NAME, e.getModuleSignature().getFirstChild().getNode().getElementType());
         assertSize(1, e.getConstraints());
         assertEquals("type t = X.t", e.getConstraints().get(0).getText());
         assertEquals("{}", e.getBody().getText());
@@ -103,7 +103,7 @@ public class ModuleParsingTest extends RmlParsingTestCase {
         RPsiInnerModule e = firstOfType(parseCode("module M: { type t; } = { type t = int; };"), RPsiInnerModule.class);
 
         assertEquals("M", e.getName());
-        assertEquals("{ type t; }", e.getModuleType().getText());
+        assertEquals("{ type t; }", e.getModuleSignature().getText());
         assertEquals("{ type t = int; }", e.getBody().getText());
     }